// apps/api/src/shared/enums/statusType.enum.ts
export const enum StatusType {
  user_status = 'user_status',
  offer_status = 'offer_status',
  item_tracking_status = 'item_tracking_status',
  flight_status = 'flight_status',
  external_flight_booking_status = 'external_flight_booking_status',
  payment_status = 'payment_status',
  notification_status = 'notification_status',
  virtual_card_status = 'virtual_card_status'
}